requirements.txt
gradio>=4.0.0
huggingface_hub>=0.24.0
llama-cpp-python>=0.2.90